What we have understood and what we have achieved. Here we have understood what is the cause of the error : "MP-Bios 8254 timer not connected to I/O APICs" Description: See these two following functions in the io_apic.c file. arch/x86/kernel/apic/io_apic.c: void __init check_timer(void) ... if (timer_irq_works()) { 	if (disable_timer_pin_1 > 0) 		clear_IO_APIC_pin(0, pin1); 	goto out; } ...
int __init timer_irq_works(void)
...
/* Let ten ticks pass... */
mdelay((10 * 1000) / HZ);
if (time_after(jiffies, t1 + 4))
   return 1;
return 0;
...
In its initialization, Linux does some bug checks, and these two functions are involved in it.
Normally, after mdelay(), jiffies must be greater than t1 + 4 (in normal nova, we have jiffies = t1 + 12).
But in our modified nova, we have jiffies = t1 + 2. That is, after mdelay, the system has registered only 2 timer interrupts. There was more than 2 timer interrupts, but only 2 were registered in the virtual linux kernel.

The problem reside in the fact that, we forced the VCpu to run a fixed maximum number of instructions (0x10000000) - Processing Element - when it does not relinquish the CPU from its own, before preempting the current PE. This cause the VCPu to register only two timer interrupts.

Solutions
One solution is to reduce the max number of instruction (by order of 0x10 for example). Doing this results now in GP Fault for unknown reason, before reaching the point of timer checking. We are suspecting two causes:
1) Performance Monitoring Interrupt Algorithm failed somewhere
    An alternative would be to introduce page fault in the second run.
2) System call in VMX non-root
   VMX - Capable system allows ring transition between user process and kernel without trapping in the hypervisor.
   Because Nova is designed to run on only VMX - capable system, We must find a mean to stop the processing element not only when there is task switch but also when there is a ring transition in the virtual machine. However, we cannot be sure now that this is the root of the problem because, until now there is is only the linux kernel that is running.

The problematic here is the sustainability of dual redundancy in real-time system. Here there is a constrain of the number of acknowledge interrupts to satisfying a certain condition.
------------------------------------------------
What we achieved:
- pushf and popf are removed from arch.hpp
- Vcpu redundant execution is achieved
	- vtlb entry is kept in the cow_elt
    - VMCS, host msr area, guest msr area, virtual APIC page are saved and checked after Processing Element execution
    - RDTSC Emulation (Simics cpu does not honor CPU_MONITOR_TRAP_FLAG in VMCS, so we find a way to circumvent it)
    - Fpu register (and MMX) saved.

+++ src/cow.cpp
@@ -34,7 +34,7 @@ Cow::~Cow() {
  * @return 
  */
 bool Cow::get_cow_list_elt(uint16 & elt_index) {
-    for (uint16 i = 0; i < NB_COW_ELT; i++) {
+    for (uint32 i = 0; i < NB_COW_ELT; i++) {
         if (!cow_list[i].used) {
             elt_index = i;
             cow_list[i].used = true;
@@ -45,7 +45,7 @@ bool Cow::get_cow_list_elt(uint16 & elt_index) {
 }
 
 bool Cow::get_new_cow_frame(uint16 & frame_index) {
-    for (uint16 i = 0; i < NB_COW_FRAME; i++) {
+    for (uint32 i = 0; i < NB_COW_FRAME; i++) {
         if (!cow_frames[i].used) {
             frame_index = i;
             cow_frames[i].used = true;
